requirements.txt
python-telegram-bot==21.10
anthropic==0.49.0
gspread==6.1.4
google-auth==2.38.0
Pillow==11.1.0
python-dotenv==1.0.1
pytest==8.3.4
